microservice-app/manage/requirements.txt
# asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.3
djangorestframework==3.13.1
pytz==2021.3
sqlparse==0.4.2
mysqlclient==2.0.1
django-cors-headers==3.5.0
pika==1.1.0
django-mysql==3.9
